Marconi.Cardano.Core.Runner

Description

Allow the execution of indexers on a Cardano node using the chain sync protocol

Runner

runIndexer :: (IsIndex (ExceptT IndexerError IO) a indexer, Closeable IO indexer, Point a ~ ChainPoint) => RunIndexerConfig a -> indexer a -> IO ()

Connect to the given socket to start a chain sync protocol and start indexing it with the given indexer.

If you want to start several indexers, use runIndexers.

Runner Config

data RunIndexerConfig event

Common configuration required to run indexers

Constructors

RunIndexerConfig (MarconiTrace IO) (RunIndexerEventPreprocessing event) RetryConfig SecurityParam NetworkId ChainPoint FilePath

runIndexerPreprocessEvent :: forall event. Lens' (RunIndexerEventPreprocessing event) (ChainSyncEvent BlockEvent -> [ProcessedInput ChainPoint event])

runIndexerExtractBlockNo :: forall event. Lens' (RunIndexerEventPreprocessing event) (event -> Maybe BlockNo)

runIndexerExtractTipDistance :: forall event. Lens' (RunIndexerEventPreprocessing event) (event -> Maybe Word)

data RunIndexerEventPreprocessing event

Runner pre-processing

Constructors

RunIndexerEventPreprocessing (ChainSyncEvent BlockEvent -> [ProcessedInput ChainPoint event]) (event -> Maybe BlockNo) (event -> Maybe Word)

runIndexerConfigTrace :: forall event. Lens' (RunIndexerConfig event) (MarconiTrace IO)

runIndexerConfigEventProcessing :: forall event event. Lens (RunIndexerConfig event) (RunIndexerConfig event) (RunIndexerEventPreprocessing event) (RunIndexerEventPreprocessing event)

runIndexerConfigRetryConfig :: forall event. Lens' (RunIndexerConfig event) RetryConfig

runIndexerConfigSecurityParam :: forall event. Lens' (RunIndexerConfig event) SecurityParam

runIndexerConfigNetworkId :: forall event. Lens' (RunIndexerConfig event) NetworkId

runIndexerConfigChainPoint :: forall event. Lens' (RunIndexerConfig event) ChainPoint

runIndexerConfigSocketPath :: forall event. Lens' (RunIndexerConfig event) FilePath

Process chainSync events

withNoPreprocessor :: RunIndexerEventPreprocessing BlockEvent

withDistancePreprocessor :: RunIndexerEventPreprocessing (WithDistance BlockEvent)

withDistanceAndTipPreprocessor :: RunIndexerEventPreprocessing TipAndBlock

Event types