Marconi.Cardano.Core.Transformer.WithSyncStats

Transformer

data WithSyncStats indexer event

A modifier that adds sync stats to the indexer

withSyncStats :: [StatsBackend] -> indexer event -> WithSyncStats indexer event

A smart constructor for WithSyncStats

Stats

data LastSyncStats

Chain synchronisation statistics measured starting from previously measured LastSyncStats.

Constructors

LastSyncStats !Word64 !Word64 ChainPoint ChainTip !(Maybe UTCTime)

syncStatsNumBlocks :: Lens' LastSyncStats Word64

syncStatsNumRollbacks :: Lens' LastSyncStats Word64

syncStatsChainSyncPoint :: Lens' LastSyncStats ChainPoint

syncStatsNodeTip :: Lens' LastSyncStats ChainTip

syncStatsLastMessageTime :: Lens' LastSyncStats (Maybe UTCTime)

emptyLastSyncStats :: LastSyncStats

Backend

data StatsBackend

The data needed to act upon stats with an arbitrary backend

Constructors

StatsBackend (LastSyncStats -> IO ()) NominalDiffTime LastSyncStats

statsBackendAction :: Lens' StatsBackend (LastSyncStats -> IO ())

statsBackendTimeBetweenActions :: Lens' StatsBackend NominalDiffTime

statsBackendState :: Lens' StatsBackend LastSyncStats