Marconi.Cardano.Core.Transformer.WithSyncStats.Backend.Printer

Documentation

data LastSyncStatsOutput

Constructors

LastSyncStatsOutput LastSyncStats (Maybe NominalDiffTime)

mkLogBackend :: MarconiTrace IO -> NominalDiffTime -> StatsBackend

Creates a simple pretty-printing backend which prints the number of processed blocks and the number of rollbacks to a given tracer.

Takes a NominalDiffTime which determines how frequently we send stats to Prometheus and a MarconiTrace IO with which it performs the tracing.