Marconi.Cardano.Core.Transformer.WithSyncStats.Backend.Prometheus

Documentation

mkPrometheusBackend :: NominalDiffTime -> IO StatsBackend

Creates a simple Prometheus backend which keeps track of the number of processed blocks and the number of rollbacks.

Takes a NominalDiffTime which determines how frequently we send stats to Prometheus.