Marconi.Cardano.Core.Types

Description

This module provides several type aliases and utility functions to deal with them.

Config for retrying

data RetryConfig

Config type for node retries

Constructors

RetryConfig

Fields

baseTimeBeforeNextRetry :: !Word64

Initial time before next retry (in seconds)

maybeMaxWaitTime :: !(Maybe Word64)

Max time before stopping retries (in seconds)

A type representing either a ChainTip or a Block, with an attached distance to the tip

data TipAndBlock

A type representing a ChainTip and maybe a Block, with an attached distance to the tip

Constructors

TipAndBlock ChainTip (Maybe (WithDistance BlockEvent))

Aliases for the current Cardano era

type TxOut = TxOut CtxTx ConwayEra

A Cardano TxOut of the current Era

Aliases to ease concept mapping between plutus types and cardano types

type TxOutRef = TxIn

A reference to a transaction output. This is a pair of a transaction reference, and an index indicating which of the outputs of that transaction we are referring to.

txOutRef :: TxId -> TxIx -> TxIn

data AnyTxBody

An existential type representing a transaction with C.TxBody era for any Cardano era.

Constructors

forall era.IsCardanoEra era => AnyTxBody BlockNo TxIndexInBlock (TxBody era)

Reexport from cardano-api-extended

data BlockEvent

Constructors

BlockEvent

Fields

blockInMode :: BlockInMode CardanoMode

epochNo :: EpochNo

blockTime :: POSIXTime

newtype SecurityParam

Constructors

SecurityParam Word64

newtype TxIndexInBlock

Constructors

TxIndexInBlock Word64